+++ app/src/main/java/models/FavoriteMovieModel.kt
@@ -4,5 +4,5 @@ import androidx.room.Entity
 import androidx.room.PrimaryKey
 
 @Entity
-data class FavoriteMovieModel (@PrimaryKey val movieId:Int, val posterPath:String) {
+data class FavoriteMovieModel (@PrimaryKey val movieId:Long, val posterPath:String) {
 }
