+++ app/src/main/java/models/MovieTrailerModel.kt
@@ -1,5 +1,5 @@
 package models
 
-data class MovieTrailerModel (val id:Int, val results:List<Result>)
+data class MovieTrailerModel (val id:Long, val results:List<Result>)
 
 data class Result (val key:String)
